requiremente.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
python-decouple==3.8
tzdata==2024.2